requirements.txt
aiohttp==3.7.4
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.10.0
bson==0.5.8
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.7
colorclass==2.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dataclasses==0.6
defusedxml==0.7.1
Django==3.2.9
django-channels-graphql-ws==0.9.0
django-common-helpers==0.9.2
django-cors-headers==3.10.0
django-cron==0.5.1
django-csp==3.7
django-environ==0.7.0
django-filter==21.1
django-graphql-jwt==0.3.4
django-graphql-social-auth==0.1.4
django-pgconnection==1.0.2
django-pgtrigger==2.4.0
django-phonenumber-field==6.0.0
django-redis==5.0.0
django-redis-cache==3.0.0
djangorestframework==3.12.4
docopt==0.6.2
graphene==2.1.9
graphene-django==2.15.0
graphene-django-cud==0.9.0
graphene-file-upload==1.3.0
graphql-core==2.2
graphql-relay==2.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.2
multidict==5.2.0
Naked==0.1.31
oauthlib==3.1.1
packaging==21.0
phonenumberslite==8.12.37
pip-upgrader==1.4.15
promise==2.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.2.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
Rx==3.2.0
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.2
terminaltables==3.1.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
zope.interface==5.4.0
